pyproject.toml
[tool.poetry]
name = "databricks-ingetion"
version = "0.1.0"
description = ""
authors = ["vikas1595 <vikasptlo7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyspark = "3.5.5"
findspark = "^1.4.2"
delta-spark = "3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^3.0"
pytest-mock = "^3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
